DrawString optimization, fixed forgotten cham code, fixed options
I could do the same to the menu, but that would mean replacing modernJson as it has no support for wchar_t, and i'd rather not. I also need to think of a way to efficiently check the items, i hope their model index is static, since that's my first idea.